refactor: 컴포넌트 및 함수명 변경
