+++ frontend/src/layouts/SidebarLayout/components/Sidebar/Sidebar.tsx
@@ -6,7 +6,7 @@ import { getGroups } from '../../../../api/group';
 import { Group } from '../../../../types/group';
 
 import SidebarGroupsMenu from '../SidebarGroupsMenu/SidebarGroupsMenu';
-import SidebarMenuModals from '../SidebarMenuModals/SidebarMenuModals';
+import SidebarModals from '../SidebarModals/SidebarModals';
 
 import SidebarMembersProfileMenu from '../SidebarMembersProfileMenu/SidebarMembersProfileMenu';
 import SidebarFeaturesMenu from '../SidebarFeaturesMenu/SidebarFeaturesMenu';
@@ -24,15 +24,15 @@ function Sidebar() {
   const navigate = useNavigate();
   const [isLoading, setIsLoading] = useState(true);
   const [groups, setGroups] = useState<Array<Group>>([]);
-  const [activeModalMenu, setActiveModalMenu] = useState<null | string>(null);
+  const [activeModal, setActiveModal] = useState<null | string>(null);
   const { groupCode } = useParams() as { groupCode: Group['code'] };
 
   const handleNavigate = (location: string) => () => {
     navigate(location);
   };
 
-  const handleActiveModalMenu = (menu: null | string) => () => {
-    setActiveModalMenu(menu);
+  const handleActiveModal = (menu: null | string) => () => {
+    setActiveModal(menu);
   };
 
   useEffect(() => {
@@ -56,15 +56,11 @@ function Sidebar() {
           <StyledLogo src={Logo} alt={Logo} onClick={handleNavigate(`/groups/${groupCode}`)} />
         </StyledLogoContainer>
 
-        <SidebarGroupsMenu
-          onClickMenu={handleActiveModalMenu}
-          groupCode={groupCode}
-          groups={groups}
-        />
+        <SidebarGroupsMenu onClickMenu={handleActiveModal} groupCode={groupCode} groups={groups} />
         <SidebarFeaturesMenu groupCode={groupCode} />
         <SidebarMembersProfileMenu />
         <StyledBottomMenu>
-          <SidebarSlackMenu onClick={handleActiveModalMenu('slack')} />
+          <SidebarSlackMenu onClick={handleActiveModal('slack')} />
           <SidebarInvitationMenu groupCode={groupCode} />
           <SidebarLogoutMenu />
         </StyledBottomMenu>
@@ -72,9 +68,9 @@ function Sidebar() {
 
       {/* TODO: 모달이 모여있음  */}
       {/* TODO: portal 사용 */}
-      <SidebarMenuModals
-        activeModalMenu={activeModalMenu}
-        closeModal={handleActiveModalMenu(null)}
+      <SidebarModals
+        activeModal={activeModal}
+        closeModal={handleActiveModal(null)}
         groupCode={groupCode}
       />
     </>
